+++ generators/common/inputs/isThisYou.js
@@ -26,8 +26,6 @@
 
 			/* istanbul ignore next */
 			if ($that.config.get('authorEmail')) {
-				logs.hintHeader($that);
-
 				prompts[0].message += $that.config.get('authorFirstName') || '';
 				prompts[0].message += ' ';
 				prompts[0].message += $that.config.get('authorLastName') || '';