+++ generators/constant/inputs/constantName.js
@@ -27,7 +27,6 @@
 
 	module.exports = class ConstantName {
 		static set($that) {
-			logs.hintHeader($that);
 			logs.hintsMultiline($that, hints);
 
 			return $that.prompt(prompts).then($response => {