+++ generators/provider/index.js
@@ -21,6 +21,7 @@
 	const colors = require('../common/colors.js');
 	const logs   = require('../common/logs.js');
 	const config = require('../common/config.js');
+	const utils = require('../common/utils.js');
 
 	const isThisYou    = require('../common/inputs/isThisYou.js');
 	const author       = require('../common/inputs/author.js');
@@ -43,7 +44,7 @@
 				{
 					type   : 'confirm',
 					name   : 'areYouReady',
-					message: 'Are you ready to set up a new provider for the ' + this.config.get('appName') + ' app ?',
+					message: 'Are you ready to set up a new provider for the ' + utils.getAppName(this) + ' app ?',
 					default: true
 				}
 			];